Brand Liaison Profile
Brand Liaison was founded in 2014 and is registered with the ROC, Ministry of Corporate Affairs, Government of India. The company provides a one-stop solution for all kinds of compliance management services, for the Indian market. We provide compliance services to Indian and foreign manufacturers such as BIS Registration, ISI Certification, BEE Registration, EPR Authorization (for E-waste and P-waste Products), WPC Approval, TEC Certification, etc., which are necessary for the notified products before importing or selling in India.